programs: Split further into arch, aur and dipm installed packages

+++ install/programs.sh
@@ -1,25 +1,30 @@
 #!/bin/sh -e
 
-sudo pacman-key --recv-key 3056513887B78AEB --keyserver keyserver.ubuntu.com
-sudo pacman-key --lsign-key 3056513887B78AEB
-sudo pacman -U --noconfirm 'https://cdn-mirror.chaotic.cx/chaotic-aur/chaotic-keyring.pkg.tar.zst'
-sudo pacman -U --noconfirm 'https://cdn-mirror.chaotic.cx/chaotic-aur/chaotic-mirrorlist.pkg.tar.zst'
+if ! grep '^Color' /etc/pacman.conf >/dev/null; then
+    sudo sed -i 's/^#Color$/Color/' /etc/pacman.conf
+fi
 
-sudo sed -i 's/^#Color$/Color/' /etc/pacman.conf
+if ! grep '\[chaotic-aur\]' /etc/pacman.conf >/dev/null; then
+    sudo pacman-key --recv-key 3056513887B78AEB --keyserver keyserver.ubuntu.com
+    sudo pacman-key --lsign-key 3056513887B78AEB
+    sudo pacman -U --noconfirm 'https://cdn-mirror.chaotic.cx/chaotic-aur/chaotic-keyring.pkg.tar.zst'
+    sudo pacman -U --noconfirm 'https://cdn-mirror.chaotic.cx/chaotic-aur/chaotic-mirrorlist.pkg.tar.zst'
 
-if ! grep '\[chaotic-aur\]' /etc/pacman.conf; then
     {
         echo '[chaotic-aur]'
         echo 'Include = /etc/pacman.d/chaotic-mirrorlist'
     } | sudo tee -a /etc/pacman.conf
+
+    sudo pacman -Syu
 fi
 
-sudo pacman -Syu
-sudo pacman -S --noconfirm yay
+if ! command -v yay >/dev/null; then
+    sudo pacman -S --noconfirm yay
+fi
 
-yay -S --noconfirm --asexplicit - <"$HOME/.config/installed-programs"
-
-curl -L https://github.com/Hejsil/dipm/releases/latest/download/dipm-x86_64-linux-musl >/tmp/dipm &&
-    chmod +x /tmp/dipm &&
-    <"$HOME/.config/dipm-installed-programs" xargs '-d\n' /tmp/dipm install &&
-    rm /tmp/dipm
+if ! command -v dipm >/dev/null; then
+    curl -L https://github.com/Hejsil/dipm/releases/latest/download/dipm-x86_64-linux-musl >/tmp/dipm &&
+        chmod +x /tmp/dipm &&
+        /tmp/dipm install dipm &&
+        rm /tmp/dipm
+fi
